Working on the saving manager, try to get the enemy data to save as well

+++ Assets/prefabs/Framework/AbilitySystem/AbilityComponent.cs
@@ -62,8 +62,8 @@ public class AbilityComponent : MonoBehaviour
         StaminaLevel = Mathf.Clamp(StaminaLevel, MaxStaminaLevel, StaminaLevel);
     }
 
-    internal int GetStaminaLevel()
+    internal float GetStaminaLevel()
     {
-        return (int)StaminaLevel;
+        return StaminaLevel;
     }
 }
